update nn/functional, test=develop
