+++ python/paddle/nn/functional/__init__.py
@@ -130,24 +130,6 @@ from .activation import sigmoid  #DEFINE_ALIAS
 # from .activation import tanh_shrink   #DEFINE_ALIAS
 # from .activation import thresholded_relu   #DEFINE_ALIAS
 from .activation import log_softmax  #DEFINE_ALIAS
-from . import extension
-__all__ += extension.__all__
-# from .extension import add_position_encoding   #DEFINE_ALIAS
-# from .extension import autoincreased_step_counter   #DEFINE_ALIAS
-# from .extension import continuous_value_model   #DEFINE_ALIAS
-# from .extension import filter_by_instag   #DEFINE_ALIAS
-# from .extension import linear_chain_crf   #DEFINE_ALIAS
-# from .extension import merge_selected_rows   #DEFINE_ALIAS
-# from .extension import multiclass_nms   #DEFINE_ALIAS
-# from .extension import polygon_box_transform   #DEFINE_ALIAS
-# from .extension import random_crop   #DEFINE_ALIAS
-from .extension import row_conv  #DEFINE_ALIAS
-# from .extension import rpn_target_assign   #DEFINE_ALIAS
-# from .extension import similarity_focus   #DEFINE_ALIAS
-# from .extension import target_assign   #DEFINE_ALIAS
-# from .extension import temporal_shift   #DEFINE_ALIAS
-# from .extension import warpctc   #DEFINE_ALIAS
-from .extension import diag_embed  #DEFINE_ALIAS
 # from .rnn import gru_unit   #DEFINE_ALIAS
 # from .rnn import lstm   #DEFINE_ALIAS
 # from .rnn import lstm_unit   #DEFINE_ALIAS
@@ -180,17 +162,3 @@ from .extension import diag_embed  #DEFINE_ALIAS
 # from .lod import dynamic_gru   #DEFINE_ALIAS
 # from .lod import dynamic_lstm   #DEFINE_ALIAS
 # from .lod import dynamic_lstmp   #DEFINE_ALIAS
-from . import common
-#__all__ += common.__all__
-# from .common import dropout   #DEFINE_ALIAS
-# from .common import embedding   #DEFINE_ALIAS
-# from .common import fc   #DEFINE_ALIAS
-# from .common import label_smooth   #DEFINE_ALIAS
-# from .common import one_hot   #DEFINE_ALIAS
-# from .common import pad   #DEFINE_ALIAS
-# from .common import pad_constant_like   #DEFINE_ALIAS
-# from .common import pad2d   #DEFINE_ALIAS
-# from .common import unfold   #DEFINE_ALIAS
-# from .common import bilinear_tensor_product   #DEFINE_ALIAS
-# from .common import assign   #DEFINE_ALIAS
-from .common import interpolate  #DEFINE_ALIAS
